refactor: presentation-layer works with domain-layer
